# sql: allow interleaving table and index data