--- a/sql/show.go
+++ b/sql/show.go
@@ -89,6 +89,26 @@ func (p *planner) ShowColumns(n *parser.ShowColumns) (planNode, error) {
 	return v, nil
 }
 
+// showCreateInterleave returns an INTERLEAVE IN PARENT clause for the specified
+// index, if applicable.
+func (p *planner) showCreateInterleave(idx *sqlbase.IndexDescriptor) (string, error) {
+	if len(idx.Interleave.Ancestors) == 0 {
+		return "", nil
+	}
+	intl := idx.Interleave
+	parentTable, err := getTableDescFromID(p.txn, intl.Ancestors[len(intl.Ancestors)-1].TableID)
+	if err != nil {
+		return "", err
+	}
+	var sharedPrefixLen int
+	for _, ancestor := range intl.Ancestors {
+		sharedPrefixLen += int(ancestor.SharedPrefixLen)
+	}
+	interleavedColumnNames := quoteNames(idx.ColumnNames[:sharedPrefixLen]...)
+	s := fmt.Sprintf(" INTERLEAVE IN PARENT %s (%s)", parentTable.Name, interleavedColumnNames)
+	return s, nil
+}
+
 // ShowCreateTable returns a CREATE TABLE statement for the specified table in
 // Traditional syntax.
 // Privileges: None.
@@ -142,11 +162,16 @@ func (p *planner) ShowCreateTable(n *parser.ShowCreateTable) (planNode, error) {
 		if len(idx.StoreColumnNames) > 0 {
 			storing = fmt.Sprintf(" STORING (%s)", quoteNames(idx.StoreColumnNames...))
 		}
-		fmt.Fprintf(&buf, ",\n\t%sINDEX %s (%s)%s",
+		interleave, err := p.showCreateInterleave(&idx)
+		if err != nil {
+			return nil, err
+		}
+		fmt.Fprintf(&buf, ",\n\t%sINDEX %s (%s)%s%s",
 			isUnique[idx.Unique],
 			quoteNames(idx.Name),
 			quoteNames(idx.ColumnNames...),
 			storing,
+			interleave,
 		)
 	}
 	for _, fam := range desc.Families {
@@ -171,6 +196,12 @@ func (p *planner) ShowCreateTable(n *parser.ShowCreateTable) (planNode, error) {
 	}
 
 	buf.WriteString("\n)")
+	interleave, err := p.showCreateInterleave(&desc.PrimaryIndex)
+	if err != nil {
+		return nil, err
+	}
+	buf.WriteString(interleave)
+
 	v.rows = append(v.rows, []parser.Datum{
 		parser.NewDString(n.Table.String()),
 		parser.NewDString(buf.String()),