--- a/sql/sqlbase/formatversion_string.go
+++ b/sql/sqlbase/formatversion_string.go
@@ -4,9 +4,9 @@ package sqlbase
 
 import "fmt"
 
-const _FormatVersion_name = "BaseFormatVersionFamilyFormatVersion"
+const _FormatVersion_name = "BaseFormatVersionFamilyFormatVersionInterleavedFormatVersion"
 
-var _FormatVersion_index = [...]uint8{0, 17, 36}
+var _FormatVersion_index = [...]uint8{0, 17, 36, 60}
 
 func (i FormatVersion) String() string {
 	i -= 1