Also changes the interleaving sentinel to encoding.descNotNull (0xfe) because
the other one could be used in keys. We want to be able to pull table and index
ids out of a key without the descriptor and it was possible for there to be
ambiguity with the previously used encoding.descNull (0xff).

Followup work:
- Allow interleaving indexes declared in a CREATE TABLE
- Support the CASCADE/RESTRICT foreign key shorthands
- Clean up the InterleavedBy back references when dropping a table/index
- Write some interleaving benchmarks
- Optimize the case where lots of interleaved data has to be scanned over (it
  currently always looks at every key.)
- Remove the SKIP DOC comments on opt_interleave in sql.y

For #2972.

Benchmarks are unchanged for non-interleaved data.
```
name                       old time/op    new time/op    delta
Insert1000_Cockroach-8       9.75ms ± 4%    9.71ms ± 5%   ~             (p=0.690 n=5+5)
Update1000_Cockroach-8       26.6ms ± 3%    26.3ms ± 5%   ~             (p=0.841 n=5+5)
Delete1000_Cockroach-8       13.1ms ± 3%    13.4ms ± 3%   ~             (p=0.421 n=5+5)
Scan1000_Cockroach-8         2.20ms ± 1%    2.20ms ± 1%   ~             (p=0.841 n=5+5)
WideTable1000_Cockroach-8     105ms ± 4%     107ms ± 6%   ~             (p=0.690 n=5+5)

name                       old alloc/op   new alloc/op   delta
Insert1000_Cockroach-8       4.25MB ± 0%    4.25MB ± 0%   ~             (p=0.111 n=4+5)
Update1000_Cockroach-8       6.42MB ± 0%    6.42MB ± 0%   ~             (p=0.841 n=5+5)
Delete1000_Cockroach-8       1.07MB ± 0%    1.07MB ± 0%   ~             (p=1.000 n=5+5)
Scan1000_Cockroach-8          317kB ± 0%     317kB ± 0%   ~             (p=0.444 n=5+5)
WideTable1000_Cockroach-8    21.3MB ± 0%    21.3MB ± 0%   ~             (p=0.222 n=5+5)

name                       old allocs/op  new allocs/op  delta
Insert1000_Cockroach-8        20.4k ± 0%     20.4k ± 0%   ~             (p=0.056 n=5+5)
Update1000_Cockroach-8        28.7k ± 0%     28.7k ± 0%   ~             (p=0.643 n=5+5)
Delete1000_Cockroach-8        12.6k ± 0%     12.6k ± 0%   ~             (p=1.000 n=5+5)
Scan1000_Cockroach-8          2.30k ± 0%     2.30k ± 0%   ~     (all samples are equal)
WideTable1000_Cockroach-8      224k ± 0%      224k ± 0%   ~             (p=0.841 n=5+5)
```
<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/cockroachdb/cockroach/7855)
<!-- Reviewable:end -->