@paperstreet 2016-07-15
I wrote all the logic test cases I could think of, but it still feels like I'm missing some. I'd really like to be more confident about this before I merge, but everything else is more-or-less done so I figured I'd get the review process started.

I wonder how hard it would be to generate a bunch of them and put them in our big test suite...

---

Review status: 0 of 14 files reviewed at latest revision, all discussions resolved, some commit checks pending.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/7855#-:-KMkEVlPfpMfU3CwUMtJ:-7450717)*
<!-- Sent from Reviewable.io -->