@paperstreet 2016-07-18
Definitely not going to be ready for this week's beta