@RaduBerinde 2016-07-18
<img class="emoji" title=":lgtm:" alt=":lgtm:" align="absmiddle" src="https://reviewable.io/lgtm.png" height="20" width="61"/>

In terms of tests, it would be nice to have cases when the first/last row fetched from KV is either the child table or the parent table (or grandparent) - just to verify that rowfetcher behaves correctly if it has to skip the first/last row.  We could do this by inserting a row in one table, rechecking some selects, repeat.

Also consider writing a standalone test that performs the same randomized workload on both a set of interleaved tables, and on a set of similar non-interleaved tables (cross-checking all results).

---

Review status: 0 of 14 files reviewed at latest revision, 10 unresolved discussions, all commit checks successful.

---

*[sql/create.go, line 339 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/7855#-KMyoB3I6u83a0gdjgjO:-KMyoB3I6u83a0gdjgjP:1375681746) ([raw file](https://github.com/cockroachdb/cockroach/blob/15f7725409df832d4d6e24732df8bf78a9aaad46/sql/create.go#L339)):*
> ```Go
> 	}
> 
> 	created, err := n.p.createDescriptor(tableKey{n.dbDesc.ID, n.n.Table.Table()}, &desc, n.n.IfNotExists)
> ```

[nit] long line

---

*[sql/create.go, line 502 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/7855#-KMyoTBUjIJ0uqnIF_2Y:-KMyoTBUjIJ0uqnIF_2Z:100445579) ([raw file](https://github.com/cockroachdb/cockroach/blob/15f7725409df832d4d6e24732df8bf78a9aaad46/sql/create.go#L502)):*
> ```Go
> }
> 
> func (p *planner) addInterleave(
> ```

a brief function comment would help

---

*[sql/create.go, line 506 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/7855#-KMyoCtiOgOCdt3Ixm0R:-KMyoCtiOgOCdt3Ixm0S:1375681746) ([raw file](https://github.com/cockroachdb/cockroach/blob/15f7725409df832d4d6e24732df8bf78a9aaad46/sql/create.go#L506)):*
> ```Go
> ) error {
> 	if interleave.DropBehavior != parser.DropDefault {
> 		return util.UnimplementedWithIssueErrorf(7854, "unsupported shorthand %s", interleave.DropBehavior)
> ```

[nit] long line

---

*[sql/create.go, line 533 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/7855#-KMyoDtPuQV2d0qbumQq:-KMyoDtPuQV2d0qbumQr:1375681746) ([raw file](https://github.com/cockroachdb/cockroach/blob/15f7725409df832d4d6e24732df8bf78a9aaad46/sql/create.go#L533)):*
> ```Go
> 			return fmt.Errorf("declared columns must match index being interleaved")
> 		}
> 		if col.Type != targetCol.Type || index.ColumnDirections[i] != parentIndex.ColumnDirections[i] {
> ```

[nit] long line

---

*[sql/create.go, line 538 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/7855#-KMyoDZ1rDZgny5Glnbo:-KMyoDZ1rDZgny5Glnbp:1375681746) ([raw file](https://github.com/cockroachdb/cockroach/blob/15f7725409df832d4d6e24732df8bf78a9aaad46/sql/create.go#L538)):*
> ```Go
> 	}
> 
> 	ancestorPrefix := append([]sqlbase.InterleaveDescriptor_Ancestor(nil), parentIndex.Interleave.Ancestors...)
> ```

[nit] long line

---

*[sql/create.go, line 587 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/7855#-KMyoVmJdA3QUD4dcPYb:-KMyoVmJdA3QUD4dcPYc:1288414015) ([raw file](https://github.com/cockroachdb/cockroach/blob/15f7725409df832d4d6e24732df8bf78a9aaad46/sql/create.go#L587)):*
> ```Go
> 
> // TODO(dan): This is similar to finalizeFKs. Consolidate them.
> func (p *planner) finalizeInterleave(
> ```

a function comment that briefly explains what the function does would help

---

*[sql/show.go, line 96 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/7855#-KMyonc434-n4wV4sNiC:-KMyonc434-n4wV4sNiD:-479785402) ([raw file](https://github.com/cockroachdb/cockroach/blob/15f7725409df832d4d6e24732df8bf78a9aaad46/sql/show.go#L96)):*
> ```Go
> // Privileges: None.
> func (p *planner) ShowCreateTable(n *parser.ShowCreateTable) (planNode, error) {
> 	showCreateInterleave := func(idx *sqlbase.IndexDescriptor) (string, error) {
> ```

this can be a separate function

---

*[sql/sqlbase/rowfetcher.go, line 201 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/7855#-KMyqw8YgRXxslglApAa:-KMyqw8YgRXxslglApAb:500653123) ([raw file](https://github.com/cockroachdb/cockroach/blob/15f7725409df832d4d6e24732df8bf78a9aaad46/sql/sqlbase/rowfetcher.go#L201)):*
> ```Go
> 		}
> 
> 		rf.keyRemainingBytes, err = rf.ReadIndexKey(rf.kv.Key)
> ```

Nice, doing this here makes more sense.

---

*[sql/sqlbase/rowfetcher.go, line 203 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/7855#-KMyqGllcENvIcNBaCxy:-KMyqGllcENvIcNBaCxz:-442163461) ([raw file](https://github.com/cockroachdb/cockroach/blob/15f7725409df832d4d6e24732df8bf78a9aaad46/sql/sqlbase/rowfetcher.go#L203)):*
> ```Go
> 		rf.keyRemainingBytes, err = rf.ReadIndexKey(rf.kv.Key)
> 		if err != nil {
> 			if _, ok := err.(KeyDescriptorMismatchError); ok {
> ```

This is the way we are "skipping" over interleaved rows that don't belong to us, correct?

It's probably not ideal to have an error involved in a "normal" hot path. There may be performance concerns, plus it would be nice to differentiate between the expected descriptors, and some arbitrary descriptor (which should be an actual error). Fine to go in as-is, but maybe leave a TODO to investigate in the future.

---

*[sql/sqlbase/table.go, line 726 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/7855#-KMyrhtTzFNbxaLIEWDc:-KMyrhtU_ScE3QUzRwL0:1639851627) ([raw file](https://github.com/cockroachdb/cockroach/blob/15f7725409df832d4d6e24732df8bf78a9aaad46/sql/sqlbase/table.go#L726)):*
> ```Go
> 			}
> 			if decodedTableID != ancestor.TableID || decodedIndexID != ancestor.IndexID {
> 				return nil, KeyDescriptorMismatchError{}
> ```

I believe that returning `KeyDescriptorMismatchError{}` hides an allocation - something must be allocated on the heap if we are to have an `error` interface point to it.

So it might be faster to use a pointer receiver when implementing `Error`, and define a "singleton" const `*KeyDescriptorMismatchError` so we always return the same pointer.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/7855#-:-KMysbUquCnRNWTijll4:28395324)*
<!-- Sent from Reviewable.io -->